+++ R/GPTscreenR-package.R
@@ -1,0 +1,12 @@
+#' @keywords internal
+"_PACKAGE"
+
+## usethis namespace: start
+#' @importFrom lemur add_message
+#' @importFrom lemur as_GPT_messages
+#' @importFrom lemur complete_GPT
+#' @importFrom lemur complete_GPT_tryCatch
+#' @importFrom lemur GPT_messages
+#' @importFrom lemur last_message
+## usethis namespace: end
+NULL