+++ R/onload.R
@@ -1,4 +1,4 @@
 .onLoad <- function(libname, pkgname) {
-  check_API_key()
-  set_model()
+  lemur::check_API_key()
+  lemur::set_model()
 }